Add files via upload
A stopwatch is a timepiece designed to measure the time interval of an event12345. It can be started and stopped by a user pressing a button or by sensors234. A stopwatch is also known as a chronometer and is used to measure fractions of time, usually short and accurate5. A stopwatch can have a hand or a digital display13.